+++ src/client/editor/tools/CustomIconTool.tsx
@@ -10,7 +10,7 @@ export default memo(function CustomIconTool() {
     <SymbolTool
       name="Custom Icon"
       order={302}
-      hotkey="extraSymbols-3"
+      hotkey="extraSymbols-4"
       defaultHidden={true}
       sample={sample}
       component={CustomIconSymbol}